+++ app/Http/Controllers/Server/SubuserController.php
@@ -79,7 +79,7 @@ class SubuserController extends Controller
             'server' => $server,
             'node' => $server->node,
             'subuser' => $subuser,
-            'permlist' => Models\Permission::list(),
+            'permlist' => Models\Permission::listPermissions(),
             'permissions' => $subuser->permissions->mapWithKeys(function ($item, $key) {
                 return [$item->permission => true];
             }),
@@ -147,7 +147,7 @@ class SubuserController extends Controller
 
         return view('server.users.new', [
             'server' => $server,
-            'permissions' => Models\Permission::list(),
+            'permissions' => Models\Permission::listPermissions(),
             'node' => $server->node,
         ]);
     }